requirements.txt
numpy==1.21.6
pandas==1.3.3
gitpython==3.1.30
tqdm==4.64.1
bytecode==0.16.0
